+++ src/database/migrations/Create/System/Module/Goods/2021_10_18_090029_create_goods_classes_table.php
@@ -18,8 +18,8 @@ use Illuminate\Support\Facades\Config;
 return new class () extends Migration {
     protected $baseTable = 'goods_classes';
     protected $hasSnowflake = false;
-		// 分片键锚定字段 仅做识别用,不参与代码逻辑（格式：*_uid，无分片则为''）
-	protected $shardKeyAnchor = '';
+    // 分片键锚定字段 仅做识别用,不参与代码逻辑（格式：*_uid，无分片则为''）
+    protected $shardKeyAnchor = '';
     protected $tableComment = '产品分类表';
 
     /**
@@ -39,7 +39,7 @@ return new class () extends Migration {
                 $table->unsignedTinyInteger('deep')->default(0)->comment('级别');
                 $table->unsignedBigInteger('revision')->default(0)->comment('乐观锁');
                 $table->unsignedTinyInteger('switch')->default(0)->comment('是否|开关 0关否1是开');
-                $table->decimal('rate', 4, 2)->default(0)->comment('分润比例%');
+                $table->unsignedBigInteger('rate')->default(0)->comment('分润比例%');
                 $table->string('goods_class_name', 64)->nullable()->comment('商品类名称');
                 $table->string('goods_class_code', 64)->nullable()->comment('商品分类逻辑名称');
                 $table->unsignedBigInteger('goods_class_picture_uid')->default(0)->comment('分类图片雪花ID(相册图片id)');
